+++ CHANGELOG.md
@@ -17,6 +17,7 @@ y este proyecto sigue [Versionado Semántico](https://semver.org/lang/es/).
 - Entidad [TypeUni](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/entity/TypeUnit.kt)
 - Entidad [Unit](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/entity/Unit.kt)
 - Repositorio [Category](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/CategoryRepository.kt) para la entidad category
+- Repositorio [Product](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/ProductRepository.kt) para la entidad product
 
 ## [0.0.1] - 2026-07-03