+++ CHANGELOG.md
@@ -18,6 +18,7 @@ y este proyecto sigue [Versionado Semántico](https://semver.org/lang/es/).
 - Entidad [Unit](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/entity/Unit.kt)
 - Repositorio [Category](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/CategoryRepository.kt) para la entidad category
 - Repositorio [Product](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/ProductRepository.kt) para la entidad product
+- Repositorio [Purchase](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/repository/PurchaseRepository.kt) para la entidad Purchase
 
 ## [0.0.1] - 2026-07-03